Bonus Feature 2 (Polling API)
Polling the API every 60 seconds and updating the UI accordingly.

Design Decision:
The polling happens in the middleware, and only a single call to the function is originated in the main page. Using the "useInterval" custom hook, it was possible to have the polling requests originate from the page, every 60 seconds. But respecting separation of concerns, it is better if the page components are responsible for the UI, rather than the logic.

Even though right now the API is static, but certain cases, where a surface is removed or updated on a future API call are also handled correctly:

- For example imagine that the user has selected a surface, and as a result, the related server is also selected. On the next API call, that surface is removed from the results, but the server still exists. In this case, not only the surface (and its details) are removed from the UI, but the previously selected server is also de-selected.

- As another example, if the user has selected a surface and a server, and on the next API call, this surface is transfered to a new server, the previously selected surface will remain selected, but the new server is selected in the UI.

- Finally, when reloading the new API data, the search term will remain unaffected. Meaning that the new results will be filtered with the same search term and displayed in the UI.